requirements.txt
absl-py==2.1.0
einops==0.7.0
hydra-core==1.3.2
immutabledict==4.2.0
librosa==0.10.1
matplotlib==3.8.4
mir-eval==0.7
note-seq==0.0.5
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.2
pretty-midi==0.2.10
pytorch-lightning==2.2.3
scipy==1.13.0
timm==0.9.16
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.3.2
transformers==4.40.1
tqdm==4.66.2
wandb==0.16.6
yacs==0.1.8